cmd-client: Refactor errors and guest logic to own files
The implementation of client is cluttered with unnecessary
noise.
